requirements.txt
aiohappyeyeballs==2.3.2
aiohttp==3.10.0
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
asgiref==3.8.1
astroid==3.2.4
async-timeout==4.0.3
attrs==23.2.0
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
dataclasses-json==0.6.7
Deprecated==1.2.14
dill==0.3.8
diskcache==5.6.3
distro==1.9.0
exceptiongroup==1.2.2
fastapi==0.112.0
filelock==3.15.4
frozenlist==1.4.1
fsspec==2024.6.1
googleapis-common-protos==1.63.2
GPUtil==1.4.0
greenlet==3.0.3
grpcio==1.65.4
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.24.5
idna==3.7
importlib_metadata==8.0.0
iniconfig==2.0.0
isort==5.13.2
Jinja2==3.1.4
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.2.11
langchain-community==0.2.10
langchain-core==0.2.25
langchain-text-splitters==0.2.2
langsmith==0.1.94
llama_cpp_python==0.2.84
MarkupSafe==2.1.5
marshmallow==3.21.3
mccabe==0.7.0
multidict==6.0.5
mypy-extensions==1.0.0
numpy==1.26.4
openai==1.37.1
opentelemetry-api==1.26.0
opentelemetry-distro==0.47b0
opentelemetry-exporter-otlp==1.26.0
opentelemetry-exporter-otlp-proto-common==1.26.0
opentelemetry-exporter-otlp-proto-grpc==1.26.0
opentelemetry-exporter-otlp-proto-http==1.26.0
opentelemetry-instrumentation==0.47b0
opentelemetry-instrumentation-asgi==0.47b0
opentelemetry-instrumentation-asyncio==0.47b0
opentelemetry-instrumentation-aws-lambda==0.47b0
opentelemetry-instrumentation-dbapi==0.47b0
opentelemetry-instrumentation-fastapi==0.47b0
opentelemetry-instrumentation-logging==0.47b0
opentelemetry-instrumentation-sqlite3==0.47b0
opentelemetry-instrumentation-threading==0.47b0
opentelemetry-instrumentation-tortoiseorm==0.47b0
opentelemetry-instrumentation-urllib==0.47b0
opentelemetry-instrumentation-wsgi==0.47b0
opentelemetry-propagator-aws-xray==1.0.1
opentelemetry-proto==1.26.0
opentelemetry-sdk==1.26.0
opentelemetry-semantic-conventions==0.47b0
opentelemetry-test-utils==0.47b0
opentelemetry-util-http==0.47b0
orjson==3.10.6
packaging==24.1
platformdirs==4.2.2
pluggy==1.5.0
prometheus_client==0.20.0
protobuf==4.25.4
psutil==6.0.0
PyAudio==0.2.14
pydantic==2.8.2
pydantic_core==2.20.1
pylint==3.2.6
pytest==8.3.2
python-dotenv==1.0.1
PyYAML==6.0.1
requests==2.32.3
sniffio==1.3.1
SpeechRecognition==3.10.4
SQLAlchemy==2.0.31
starlette==0.37.2
tenacity==8.5.0
tomli==2.0.1
tomlkit==0.13.0
tqdm==4.66.4
typing-inspect==0.9.0
typing_extensions==4.12.2
urllib3==2.2.2
uvicorn==0.30.5
wrapt==1.16.0
yarl==1.9.4
zipp==3.19.2
